+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/security/SecurityConfig.java
@@ -123,6 +123,8 @@ public class SecurityConfig {
                         "/single-level-bom-as-planned/**",
                         "/edc/**",
                         "/erp-adapter/**",
+                        "/irs/root-grants/**",
+                        "/irs/partner-grants/**",
                         "/irs/requests/**",
                         "/parttypeinformation/**",
                         "/files/**",